Classe NotaDao e interface

+++ src/main/java/com/accenture/treinamento/projeto/portal/model/NotaBean.java
@@ -2,10 +2,10 @@ package com.accenture.treinamento.projeto.portal.model;
 
 public class NotaBean {
 	private Integer id_nota;
-	private double nota1;
-    private double nota2;
-    private double media;
-    private double mediafinal;
+	private float nota1;
+    private float nota2;
+    private float media;
+    private float mediafinal;
     private DisciplinaBean disciplica;
     
     public NotaBean(){
@@ -18,28 +18,28 @@ public class NotaBean {
 	public void setId_nota(Integer id_nota) {
 		this.id_nota = id_nota;
 	}
-	public double getNota1() {
+	public float getNota1() {
 		return nota1;
 	}
-	public void setNota1(double nota1) {
+	public void setNota1(float nota1) {
 		this.nota1 = nota1;
 	}
-	public double getNota2() {
+	public float getNota2() {
 		return nota2;
 	}
-	public void setNota2(double nota2) {
+	public void setNota2(float nota2) {
 		this.nota2 = nota2;
 	}
-	public double getMedia() {
+	public float getMedia() {
 		return media;
 	}
-	public void setMedia(double media) {
+	public void setMedia(float media) {
 		this.media = media;
 	}
-	public double getMediafinal() {
+	public float getMediafinal() {
 		return mediafinal;
 	}
-	public void setMediafinal(double mediafinal) {
+	public void setMediafinal(float mediafinal) {
 		this.mediafinal = mediafinal;
 	}
 	public DisciplinaBean getDisciplica() {
